# internal/server/static/openapi/local-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CSGHub Lite Local API",
    "version": "0.1.0",
    "description": "Structured documentation for the local `/api/*` and `/v1/*` endpoints served by `csghub-lite`.\n\nMaintenance note: whenever you add, remove, or rename an `/api/*` or `/v1/*` route in `internal/server/routes.go`, update this file and run the OpenAPI sync test before merging."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current csghub-lite server"
    },
    {
      "url": "http://localhost:11435",
      "description": "Default local server"
    }
  ],
...
  "paths": {
...
    "/api/tags": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List available models",
        "description": "Returns local, CSGHub cloud, and selected third-party provider models. Use `provider` to filter by a model provider such as `local`, `csghub`, or a third-party provider name. Third-party provider filters return only models selected through `/api/tags/manage`.",
        "operationId": "getTags",
        "parameters": [
          {
            "name": "refresh",
            "in": "query",
            "description": "Force a refresh of the cloud model catalog before returning results.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "description": "Filter models by normalized provider name. Local models use `local`; all CSGHub cloud models use `csghub`; third-party provider models use the configured provider name, not the generated provider ID.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Optional csghub-lite category key from `/api/pipeline-tags`. When provided, only models whose pipeline tag belongs to the category are returned.",
            "schema": {
              "type": "string",
              "enum": [
                "language_model",
                "image_generation",
                "video_generation",
                "text_to_speech",
                "speech_recognition"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Available models.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/tags/manage": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List provider catalog models",
        "description": "Returns all currently available models from a configured third-party provider. This catalog is used to choose which provider models appear in `/api/tags?provider=...`.",
        "operationId": "listProviderTagCatalog",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": true,
            "description": "Third-party provider ID or configured provider name.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Optional csghub-lite category key from `/api/pipeline-tags`. When omitted, all provider catalog models are returned.",
            "schema": {
              "type": "string",
              "enum": [
                "language_model",
                "image_generation",
                "video_generation",
                "text_to_speech",
                "speech_recognition"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider catalog models.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Select a provider model",
        "operationId": "addProviderTagModel",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": true,
            "description": "Third-party provider ID or configured provider name.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderTagModelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model was already selected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelInfo"
                }
              }
            }
          },
          "201": {
            "description": "Model selected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelInfo"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "Models"
        ],
        "summary": "Replace selected provider models",
        "operationId": "replaceProviderTagModels",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": true,
            "description": "Third-party provider ID or configured provider name.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Optional csghub-lite category key from `/api/pipeline-tags`. When supplied, the response is filtered to selected models in that category.",
            "schema": {
              "type": "string",
              "enum": [
                "language_model",
                "image_generation",
                "video_generation",
                "text_to_speech",
                "speech_recognition"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderTagModelsReplaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Selected provider models.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "Models"
        ],
        "summary": "Update selected provider model metadata",
        "description": "Updates metadata for a selected third-party provider model and can change the client-visible model ID. The original provider model ID is preserved internally for upstream inference calls. Use an empty string to clear an overridden display name or description.",
        "operationId": "updateProviderTagModel",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": true,
            "description": "Third-party provider ID or configured provider name.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "description": "Current client-visible provider model ID to update. Required when changing the model ID; the request body `model` value is the new client-visible ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderTagModelUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated provider model metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelInfo"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      },
      "delete": {
        "tags": [
          "Models"
        ],
        "summary": "Unselect a provider model",
        "operationId": "deleteProviderTagModel",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": true,
            "description": "Third-party provider ID or configured provider name.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": true,
            "description": "Provider model ID to remove from the selected list.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model unselected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/models/search": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Search local downloaded models",
        "operationId": "searchLocalModels",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Case-insensitive search text matched against model metadata.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Filter by local model format, such as `gguf`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pipeline_tag",
            "in": "query",
            "description": "Filter by detected pipeline type, such as `text-generation` or `image-text-to-text`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return. Defaults to `20`, capped at `100`.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Zero-based offset into the filtered result set.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching local models.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalModelSearchResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/models/{namespace}/{name}/manifest": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Get a local model download manifest",
        "operationId": "getModelManifest",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Local model metadata with downloadable files.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelManifestResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/models/upload": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Upload a local model",
        "description": "Imports a local model from multipart form data. Use `mode=archive` for a single `.zip`, `.tar`, `.tar.gz`, or `.tgz` file; use `mode=files` or `mode=directory` with one or more files and matching `paths` values.",
        "operationId": "uploadModel",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "files"
                ],
                "properties": {
                  "model": {
                    "type": "string",
                    "description": "Optional target model ID in `namespace/name` form. Defaults to `local/<filename>` when omitted."
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "archive",
                      "directory",
                      "files"
                    ],
                    "default": "files"
                  },
                  "overwrite": {
                    "type": "boolean",
                    "default": false
                  },
                  "paths": {
                    "type": "array",
                    "description": "Relative paths corresponding to uploaded files. Folder uploads should provide browser-relative paths.",
                    "items": {
                      "type": "string"
                    }
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The imported local model.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelUploadResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/models/{namespace}/{name}/files/{path...}": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Download a cached local model file",
        "description": "Downloads a cached model file and supports HTTP Range requests for resumable downloads. HEAD requests are also supported.",
        "operationId": "downloadModelFile",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path...",
            "in": "path",
            "required": true,
            "description": "Slash-separated relative path to the cached file within the model directory.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested file.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "206": {
            "description": "Partial file content for a range request.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/show": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Show model details",
        "operationId": "showModel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShowRequest"
              },
              "example": {
                "model": "Qwen/Qwen3-0.6B-GGUF"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/pull": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Download a model",
        "description": "Streams model download progress as Server-Sent Events.",
        "operationId": "pullModel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PullRequest"
              },
              "example": {
                "model": "Qwen/Qwen3-0.6B-GGUF"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Streaming download progress.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/EventStream"
                },
                "example": "data: {\"status\":\"pulling Qwen/Qwen3-0.6B-GGUF\"}\n\ndata: {\"status\":\"downloading model.gguf\",\"digest\":\"model.gguf\",\"total\":123456,\"completed\":4096}\n\ndata: {\"status\":\"success\"}\n\n"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/load": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Eagerly load a model",
        "description": "Loads a model into memory and optionally streams progress while conversions are running.",
        "operationId": "loadModel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadRequest"
              },
              "example": {
                "model": "Qwen/Qwen3-0.6B-GGUF",
                "stream": false,
                "num_ctx": 131072,
                "num_parallel": 1,
                "n_gpu_layers": 40,
                "cache_type_k": "q8_0",
                "cache_type_v": "q8_0",
                "dtype": "q8_0"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Load completed or streaming progress returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadResponse"
                },
                "example": {
                  "status": "ready"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/EventStream"
                },
                "example": "data: {\"status\":\"loading Qwen/Qwen3-0.6B-GGUF\"}\n\ndata: {\"status\":\"converting\",\"step\":\"quantizing\",\"current\":1,\"total\":3}\n\ndata: {\"status\":\"ready\"}\n\n"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/stop": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Stop a running model",
        "operationId": "stopModel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StopRequest"
              },
              "example": {
                "model": "Qwen/Qwen3-0.6B-GGUF"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model stopped.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                },
                "example": {
                  "status": "stopped"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/delete": {
      "delete": {
        "tags": [
          "Models"
        ],
        "summary": "Delete a local model",
        "operationId": "deleteModel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRequest"
              },
              "example": {
                "model": "Qwen/Qwen3-0.6B-GGUF"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                },
                "example": {
                  "status": "deleted"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/generate": {
      "post": {
        "tags": [
          "Inference"
        ],
        "summary": "Generate text",
        "description": "Generates text from a prompt. By default the endpoint streams tokens as Server-Sent Events unless `stream` is set to `false`.",
        "operationId": "generateText",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              },
              "example": {
                "model": "Qwen/Qwen3-0.6B-GGUF",
                "prompt": "Write a haiku about local inference.",
                "stream": false,
                "options": {
                  "temperature": 0.7,
                  "max_tokens": 128
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated text or a streaming event sequence.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/EventStream"
                },
                "example": "data: {\"model\":\"Qwen/Qwen3-0.6B-GGUF\",\"response\":\"Local\",\"done\":false,\"created_at\":\"2026-03-31T12:00:00Z\"}\n\ndata: {\"model\":\"Qwen/Qwen3-0.6B-GGUF\",\"response\":\" inference\",\"done\":false,\"created_at\":\"2026-03-31T12:00:00Z\"}\n\ndata: {\"model\":\"Qwen/Qwen3-0.6B-GGUF\",\"response\":\"\",\"done\":true,\"created_at\":\"2026-03-31T12:00:01Z\"}\n\n"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/datasets": {
      "get": {
        "tags": [
          "Datasets"
        ],
        "summary": "List local datasets",
        "operationId": "getDatasets",
        "responses": {
          "200": {
            "description": "Available local datasets.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetTagsResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/datasets/{namespace}/{name}/manifest": {
      "get": {
        "tags": [
          "Datasets"
        ],
        "summary": "Get a local dataset download manifest",
        "operationId": "getDatasetManifest",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Local dataset metadata with downloadable files.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetManifestResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "The request body, query parameters, or path parameters were invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "The requested model, dataset, app, or shell session was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Conflict": {
        "description": "The request conflicts with the current AI App state.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AIAppInfo"
            }
          }
        }
      },
      "InternalError": {
        "description": "An internal server error occurred.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "TagsResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "models"
        ],
        "properties": {
          "models": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelInfo"
            }
          }
        }
      },
...
    }
  }
}